Refatora camada service genero

+++ src/main/java/com/afzdev/demo/livraria/service/impl/GeneroServiceImpl.java
@@ -1,15 +1,21 @@
 package com.afzdev.demo.livraria.service.impl;
 
+import com.afzdev.demo.livraria.dto.AutorDTO;
 import com.afzdev.demo.livraria.dto.GeneroDTO;
+import com.afzdev.demo.livraria.entities.Autor;
+import com.afzdev.demo.livraria.entities.Genero;
+import com.afzdev.demo.livraria.exceptions.DataBaseException;
+import com.afzdev.demo.livraria.exceptions.ResourceNotFoundException;
 import com.afzdev.demo.livraria.mapper.GeneroMapper;
 import com.afzdev.demo.livraria.repository.GeneroRepository;
 import com.afzdev.demo.livraria.service.GeneroService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class GeneroServiceImpl implements GeneroService {
@@ -30,12 +36,12 @@ public class GeneroServiceImpl implements GeneroService {
     }
 
     @Override
-    @Transactional(readOnly = true)
-    public Optional<GeneroDTO> buscarPorId(Long id) {
-        return Optional.of(generoMapper.toDTO(generoRepository.findById(id)
-                .orElseThrow(()-> new RuntimeException("Não foi encontrado Genero com id : "+id))));
-
+    public GeneroDTO buscarPorId(Long id) {
+        return generoRepository.findById(id)
+                .map(generoMapper::toDTO)
+                .orElseThrow(()-> new ResourceNotFoundException("Não foi encontrado Autor com id : "+id));
     }
+
 
     @Override
     @Transactional
@@ -45,13 +51,34 @@ public class GeneroServiceImpl implements GeneroService {
 
 
     @Override
-    @Transactional
-    public void excluir(Long id) {
-        generoRepository.deleteById(id);
+    public GeneroDTO atualizar(Long id,GeneroDTO generoDTO) {
+        try {
+            Genero genero = generoRepository.getReferenceById(id);
+            updateToEntity(generoDTO,genero);
+            return generoMapper.toDTO(generoRepository.save(genero));
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Não foi encontrado Genero com id : "+id);
+        }
     }
 
+
     @Override
-    public boolean existePorId(Long id) {
-        return generoRepository.existsById(id);
+    @Transactional
+    public void excluir(Long id) {
+        try{
+            if (!generoRepository.existsById(id)){
+                throw new ResourceNotFoundException("Não foi encontrado Genero com "+id+" ao tentar excluir");
+            }
+            generoRepository.deleteById(id);
+        }catch (
+                DataIntegrityViolationException e){
+            throw new DataBaseException("Falha de integridade referencial");
+        }
+
+    }
+
+    //auxiliares
+    private void updateToEntity(GeneroDTO dto, Genero entity){
+        entity.setNome(dto.getNome());
     }
 }
